requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
astor==0.8.1
async-timeout==4.0.3
attrdict==2.0.1
Babel==2.14.0
bce-python-sdk==0.9.4
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.8
decorator==5.1.1
deep-sort-realtime==1.3.2
defusedxml==0.7.1
Django==5.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.2
flask-babel==4.0.0
fonttools==4.49.0
fsspec==2024.2.0
future==1.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
ImageHash==4.3.1
imageio==2.34.0
imgaug==0.4.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
lapx==0.5.5
lazy_loader==0.3
lmdb==1.4.1
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
paddle-bfloat==0.1.7
paddleocr==2.7.0.3
paddlepaddle==2.5.1
pandas==2.2.1
pdf2docx==0.5.8
pillow==10.2.0
premailer==3.10.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.2
pydantic_core==2.16.3
PyJWT==2.8.0
PyMuPDF==1.20.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==1.1.0
python3-openid==3.2.0
pytz==2024.1
PyWavelets==1.5.0
PyYAML==6.0.1
rapidfuzz==3.6.1
rarfile==4.1
redis==5.0.2
requests==2.31.0
requests-oauthlib==1.3.1
scikit-image==0.22.0
scipy==1.12.0
seaborn==0.13.2
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.5
sqlparse==0.4.4
supervision==0.18.0
sympy==1.12
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2024.2.12
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
trinetraservices==0.1
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.26
urllib3==2.2.1
visualdl==2.5.3
Werkzeug==3.0.1
